Updating existing integration tests for static typing

+++ src/commonMain/kotlin/indexer/codegen/IndexerParentSubaccountResponse.kt
@@ -25,7 +25,7 @@ import kotlinx.serialization.Serializable
 data class IndexerParentSubaccountResponse(
 
     val address: kotlin.String? = null,
-    val parentSubaccountNumber: kotlin.Double? = null,
+    val parentSubaccountNumber: kotlin.Int? = null,
     val equity: kotlin.String? = null,
     val freeCollateral: kotlin.String? = null,
     val childSubaccounts: kotlin.Array<IndexerSubaccountResponseObject>? = null
